# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ZSTIO-CMS API DOCS",
    "version": "1.0"
  },
  "paths": {
...
    "/api/weather/manage": {
      "get": {
        "summary": "Get weather settings",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "weather": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string"
                        },
                        "show": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Update weather settings",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "showWeather": {
                    "type": "boolean"
                  },
                  "weatherApi": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Weather settings updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "weather": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string"
                        },
                        "show": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "default": {
        "summary": "Method not allowed",
        "responses": {
          "405": {
            "description": "Method not allowed"
          }
        }
      }
    }
  },
...
}